requirements.txt
Flask==1.1.2
fuzzywuzzy==0.18.0
pysqlite3==0.4.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-engineio==3.12.1
requests==2.23.0
virtualenv==20.0.18
yelp==1.0.2
boto3
gunicorn
sendgrid
gunicorn